% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET.R
\name{aqiGet}
\alias{aqiGet}
\title{aqiGet 
Air Quality Index for major Swedish cities}
\usage{
aqiGet()
}
\value{
List dataframes for AQI
}
\description{
A Shiny package over air quality for 6 major cities in Sweden. Particulate Matter 10 (PM10) is plotted versus one year if possible.
}
\references{
https://www.naturvardsverket.se/amnesomraden/luft/statistik--utslapp-och-halter/luftkvaliteten-i-realtid-och-preliminar-statistik/webbtjanster-luftkvalitetsdata
}
