% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAPI.R
\name{check_fuel_region_time}
\alias{check_fuel_region_time}
\title{Get the electricity sources used between specified datetimes for a specified region.}
\usage{
check_fuel_region_time(region = "Scotland", from = "now",
  to = "tomorrow")
}
\arguments{
\item{region}{A number (regionid) or a string (region short name). If nothing specified, the default is "Scotland". See \code{regions()} to know the admitted values.}

\item{from}{A string with shape yyyy-mm-ddTHH:mmZ, representing the start datetime.}

\item{to}{A string with shape yyyy-mm-ddTHH:mmZ, representing the end datetime. The end date time can be max. 1 month after \code{from} date.}
}
\value{
A data frame with two columns, fuel and percentage. Which represents the percentage of the fuels used to generate electricity in the specified datetime.
}
\description{
Get the electricity sources used between specified datetimes for a specified region.
}
\examples{
check_fuel_region_time()
check_fuel_region_time("london", from = "2018-11-16T06:00Z", to = "2018-12-15T06:00Z")
}
