% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronaApi.R
\docType{class}
\name{coronaApi-class}
\alias{coronaApi-class}
\alias{coronaApi}
\title{this package is used to retrieve data for covid 19}
\description{
this package is used to retrieve data for covid 19
}
\section{Fields}{

\describe{
\item{\code{url}}{character.}

\item{\code{response_list}}{list.}

\item{\code{countries}}{list.}
}}


