% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuotesGenerator.R
\name{fetchAllQuotes}
\alias{fetchAllQuotes}
\title{All quotes function}
\usage{
fetchAllQuotes()
}
\value{
the list containing id,author's name and quote
}
\description{
This function generates all quotes
}
