% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuotesGenerator.R
\name{fetchQuoteByAuther}
\alias{fetchQuoteByAuther}
\title{All quotes by author name function}
\usage{
fetchQuoteByAuther(author)
}
\arguments{
\item{author}{name}
}
\value{
the list containing id,author's name and quote
}
\description{
This function generates all quotes of an author
}
