% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuotesGenerator.R
\name{fetchRandomQuote}
\alias{fetchRandomQuote}
\title{Random quotes function}
\usage{
fetchRandomQuote()
}
\value{
the list containing id,author's name and quote
}
\description{
This function generates random quote
}
