% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmObjects.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Get request for boundary box of the element}
\usage{
getBoundingBox(city)
}
\arguments{
\item{city}{}
}
\value{
matrix with coordinates
}
\description{
Get request for boundary box of the element
}
