% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmObjects.R
\name{getElements}
\alias{getElements}
\title{get request for element from overpass API}
\usage{
getElements(bbox, key, value)
}
\arguments{
\item{bbox}{boundaries}

\item{key}{ex. amenity}

\item{value}{ex. atm}
}
\value{
data frame of coordinates
}
\description{
get request for element from overpass API
}
