% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyFigures.R
\name{getKeyFigures}
\alias{getKeyFigures}
\title{Get Key figures}
\usage{
getKeyFigures(search_item)
}
\arguments{
\item{search_item}{search string}
}
\value{

}
\description{
get data from the API
}
\examples{

}
