% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{is_valid_year}
\alias{is_valid_year}
\title{Function to check if a string is a valid year}
\usage{
is_valid_year(input_string)
}
\arguments{
\item{input_string}{input year string}
}
\value{
bool to show it's valid or not
}
\description{
Function to check if a string is a valid year
}
