% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{kolada_municipality}
\alias{kolada_municipality}
\title{Get kolada municipality information(id,title and type) in JSON object}
\usage{
kolada_municipality(search_title = "")
}
\arguments{
\item{search_title}{keywords to search in title field,optional}
}
\value{
kolada municipality information in JSON format, if error happens, return NULL
}
\description{
Get kolada municipality information(id,title and type) in JSON object
}
