% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{kolada_municipality_kpi}
\alias{kolada_municipality_kpi}
\title{Get kolada municipality kpi JSON object}
\usage{
kolada_municipality_kpi(kpi, municipality, year)
}
\arguments{
\item{kpi}{kpi to search}

\item{municipality}{city id to search}

\item{year}{years to search}
}
\value{
kolada municipality kpis in JSON format, if error happens, return NULL
}
\description{
Get kolada municipality kpi JSON object
}
