% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{kolada_municipality_kpi_groups}
\alias{kolada_municipality_kpi_groups}
\title{Get kolada municipality kpi groups  JSON object}
\usage{
kolada_municipality_kpi_groups(search_title = "")
}
\arguments{
\item{search_title}{keywords to search in title field}
}
\value{
kolada municipality kpi groups in data frame, if error happens, return NULL
}
\description{
Get kolada municipality kpi groups  JSON object
}
