% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Municipality_Statistics.R
\name{Municipality}
\alias{Municipality}
\title{MunicipalityInput}
\usage{
Municipality(MunicipalityInput)
}
\arguments{
\item{MunicipalityInput}{takes in this argument and plots a graph.}
}
\value{
returns a plot with votecount
}
\description{
MunicipalityInput
}
\examples{


{

 Municipality(MunicipalityInput =   "Stockholms län")
}
}
