% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipals.R
\docType{class}
\name{municipals-class}
\alias{municipals-class}
\alias{municipals}
\title{municipals RC class}
\value{
information about municipilities
}
\description{
This is a RC class for handle municipal data
}
\section{Fields}{

\describe{
\item{\code{base_URL}}{character.}

\item{\code{all_data}}{data.frame.}

\item{\code{all_council_data}}{data.frame.}

\item{\code{all_municipal_data}}{data.frame.}

\item{\code{selected_municipal}}{character.}
}}


