% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmObjects.R
\docType{class}
\name{osmObjects-class}
\alias{osmObjects-class}
\alias{osmObjects}
\title{gets Coordinates of objects in the specified Location}
\value{
osmObjects class
}
\description{
gets Coordinates of objects in the specified Location
}
\section{Fields}{

\describe{
\item{\code{Location}}{character.}

\item{\code{OsmData}}{list.}

\item{\code{Coordinates}}{.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(city, object)}}{constructor object = atm, bank, pharmacy, hospital, supermarket, fuel}

\item{\code{plot()}}{Plots map with marked objects}

\item{\code{print()}}{prints coordinates of requested objects}
}}

