% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MunicipalityPercentage.R
\name{PercentageCalc}
\alias{PercentageCalc}
\title{To check the peecentage of the election results in the municipalities}
\usage{
PercentageCalc(PercentageInput)
}
\arguments{
\item{PercentageInput}{this imports the municipality which from which each party has received how man votes}
}
\value{
returns a graph showing the plot between the percentage of party votes and the municipality
}
\description{
To check the peecentage of the election results in the municipalities
}
\references{
"google.com"
}
