% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuotesGenerator.R
\docType{data}
\name{Quotes}
\alias{Quotes}
\alias{baseUrl}
\title{Lab5: Quotes generator API}
\format{
An object of class \code{character} of length 1.
}
\usage{
baseUrl
}
\description{
This package generates quotes of different authors
}
\keyword{datasets}
