% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{readJSONFromUrl}
\alias{readJSONFromUrl}
\title{Get Json R object from a URL}
\usage{
readJSONFromUrl(url)
}
\arguments{
\item{url}{URL}
}
\value{
Json object
}
\description{
Get Json R object from a URL
}
