% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAPI.R
\name{regions}
\alias{regions}
\title{A function to know the regions admitted by check_fuel_region_time().}
\usage{
regions()
}
\value{
A data frame with two columns, regionid and shortname.
}
\description{
A function to know the regions admitted by check_fuel_region_time().
}
