% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchArea.R
\name{searchArea}
\alias{searchArea}
\title{searchArea}
\usage{
searchArea(search_area)
}
\arguments{
\item{search_area}{Search text about search municipal area name}
}
\value{
Area Information about Municipility
}
\description{
get data from the API.
}
\details{
Search Area Information
}
