% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{qr app}
\usage{
server(input, output)
}
\arguments{
\item{input}{from user}

\item{output}{wr we get}
}
\value{
app
}
\description{
bankend part of app where all executions happen.
}
