\name{Team05Lab05-package}
\alias{Team05Lab05-package}
\alias{Team05Lab05}
\docType{package}
\title{
\packageTitle{Team05Lab05}
}
\description{
\packageDescription{Team05Lab05}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Team05Lab05}
\packageIndices{Team05Lab05}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Team05Lab05}

Maintainer: \packageMaintainer{Team05Lab05}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
