% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team05Lab05.R
\docType{package}
\name{Team05Lab05}
\alias{Team05Lab05}
\alias{Team05Lab05-package}
\title{A package for GETting information from https://carbonintensity.org.uk/.}
\description{
This package contains three functions \code{regions()}, \code{check_fuel_region_time()} and \code{pie_API_2()}
}
\details{
This package calls data from the Carbon Intensity API and does some manipulations to organize the data and plot it in a pie chart.
}
\references{
More information on Carbon Intensity UK API \href{https://carbonintensity.org.uk/}{here}.
}
\author{
\strong{Maintainer}: Jose Mendez \email{josme478@student.liu.se}

Authors:
\itemize{
  \item Sara Salahshour \email{sarsa293@student.liu.se}
  \item Atieh Khaleghi \email{atikh281@student.liu.se}
}

}
