% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_qr}
\alias{fetch_qr}
\title{Fetch_qr}
\usage{
fetch_qr(name, size)
}
\arguments{
\item{name}{a string}

\item{size}{a numeric}
}
\value{
A Vector
}
\description{
Fetch_qr
}
\examples{
fetch_qr("hamza",200*200)
}
