% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmuni.r
\name{getColumns}
\alias{getColumns}
\title{To check the average percentage election result from each section of the municipality}
\usage{
getColumns()
}
\value{
returns a graph showing the plot between the percentage of party votes and the municipality
}
\description{
To check the average percentage election result from each section of the municipality
}
\references{
"google.com"
}
