% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyFigures.R
\name{getKeyFigures}
\alias{getKeyFigures}
\title{getKeyFigures}
\usage{
getKeyFigures(search_item)
}
\arguments{
\item{search_item}{search string}
}
\value{
Test getKeyFigures func
}
\description{
get data from the API.
}
\details{
Get Key figures
}
