% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab05-package.R
\docType{package}
\name{lab05-package}
\alias{lab05-package}
\alias{lab05}
\title{Lab 5 of Advanced Programming in R}
\description{
This package include following functions\cr
1. kolada_municipality_kpi_groups \cr
2. kolada_municipality\cr
3.kolada_municipality_kpi\cr
\cr
}
